mv studies query funcs to methods
